Chandrakirti-chariot-swami-sarvapriyananda-2023
This essay was an assignment of Swami Sarvapriyananda as a Nagral Fellow in Harvard Divinity School in 2019. This was his term paper assignment of Indo Tibetian Madhyamaka Course, taught by Professor Jay L. Garfield. This seminal essay is now the final chapter in Swami's 2025 book on "Fullness & Emptiness: Vedanta and Buddhism". In 2023, this essay was originally published in 3 parts in January 2023, February 2023 and March 2023 of the monthly magazine The Vedanta Kesari. Fun Anecdote - Swami has told in his talk's that professor Garfield made him re-write this essay 3 times and also reviewed it as "it is not a temple talk, give proper logic - for and against - for every arguments".